# docs/swagger.yaml
basePath: /v1
definitions:
...
  Road-To-Destination-BE_module_trip_model_response.TripResponse:
    properties:
      createdTime:
        type: string
      endTime:
        type: string
      id:
        format: uuid
        type: string
      memberLimit:
        example: 15
        type: integer
      myRole:
        example: leader
        type: string
      name:
        type: string
      note:
        type: string
      ownerId:
        format: uuid
        type: string
      startTime:
        type: string
      status:
        example: planning
        type: string
      totalDistance:
        type: number
      tripType:
        example: bronze
        type: string
      updatedTime:
        type: string
      visibility:
        type: boolean
    type: object
  enum.DestinationStatus:
    enum:
    - 0
    - 1
    - 2
    type: integer
    x-enum-varnames:
    - Editing
    - Confirm
    - Deleted
  model.Compound:
    properties:
      commune:
        type: string
      district:
        type: string
      province:
        type: string
    type: object
  model.Destination:
    properties:
      arriveTime:
        type: string
      createdTime:
        type: string
      id:
        format: uuid
        type: string
      lat:
        type: number
      lng:
        type: number
      location:
        $ref: '#/definitions/model.Location'
      locationId:
        type: string
      name:
        type: string
      status:
        $ref: '#/definitions/enum.DestinationStatus'
      stayOverTime:
        description: đơn vị là phút
        type: integer
      updatedTime:
        type: string
    type: object
...
  model.Location:
    properties:
      address:
        type: string
      compound:
        $ref: '#/definitions/model.Compound'
      createdTime:
        type: string
      formattedAddress:
        type: string
      id:
        format: uuid
        type: string
      isVerified:
        type: boolean
      lat:
        type: number
      lng:
        type: number
      name:
        type: string
      placeId:
        type: string
      placeType:
        $ref: '#/definitions/model.PlaceType'
      placeTypeId:
        type: string
      updatedTime:
        type: string
    type: object
  model.PlaceType:
    properties:
      createdTime:
        type: string
      id:
        format: uuid
        type: string
      typeName:
        type: string
      updatedTime:
        type: string
    type: object
...
  request.ForkTripRequest:
    properties:
      endTime:
        type: string
      excludeUserIds:
        items:
          type: string
        type: array
      name:
        maxLength: 255
        minLength: 1
        type: string
      note:
        maxLength: 4000
        type: string
      startTime:
        type: string
      tripType:
        example: bronze
        type: string
      visibility:
        type: boolean
    required:
    - excludeUserIds
    type: object
...
  request.UpdateDestinationRequest:
    properties:
      arriveTime:
        type: string
      lat:
        type: number
      lng:
        type: number
      name:
        type: string
      status:
        $ref: '#/definitions/enum.DestinationStatus'
      stayOverTime:
        type: integer
    type: object
...
  request.UpdateTripRequest:
    properties:
      endTime:
        type: string
      name:
        maxLength: 255
        minLength: 1
        type: string
      note:
        maxLength: 4000
        type: string
      startTime:
        type: string
      visibility:
        type: boolean
    type: object
...
  response.TripMemberResponse:
    properties:
      invitorName:
        type: string
      joinedAt:
        type: string
      nickname:
        type: string
      role:
        example: member
        type: string
      status:
        example: active
        type: string
      tripId:
        format: uuid
        type: string
      tripName:
        type: string
      userId:
        format: uuid
        type: string
      username:
        type: string
    type: object
  share.ErrorResponse:
    properties:
      errors:
        additionalProperties:
          type: string
        type: object
      message:
        example: invalid input
        type: string
      status:
        example: 400
        type: integer
    type: object
info:
  contact: {}
  description: 'Backend API for Road2D (team Vandra): collaborative multi-branch ride
    planning and group-ride safety in Vietnam. JWT auth, Goong Maps playground, trip
    graph models.'
  title: Road2D
  version: "1.0"
paths:
...
  /planning/destination/{destinationId}:
    delete:
      description: Deletes a destination that is still editing. A pin that sits only
        on draft branches is removed immediately. A pin on a route branch is stripped
        first; the delete is rejected when that would disconnect the graph or when
        a trip that is no longer planning still uses the pin on a route. Travels that
        pointed at the pin are removed with the row.
      parameters:
      - description: Destination UUID
        format: uuid
        in: path
        name: destinationId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: Destination deleted
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/share.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/share.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/share.ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/share.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Delete destination
      tags:
      - planning
    get:
      description: Returns a destination. When locationId is set, lat/lng are locked
        to that location.
      parameters:
      - description: Destination UUID
        format: uuid
        in: path
        name: destinationId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Destination'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/share.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/share.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/share.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get destination
      tags:
      - planning
    put:
      consumes:
      - application/json
      description: Updates name, arrive time, stay, and status while the destination
        is editing. lat/lng are applied only when locationId is empty; a forked or
        verified pin keeps the linked location's coordinates.
      parameters:
      - description: Destination UUID
        format: uuid
        in: path
        name: destinationId
        required: true
        type: string
      - description: Fields to store
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/request.UpdateDestinationRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Destination'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/share.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/share.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/share.ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/share.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update destination
      tags:
      - planning
...
  /planning/location/{locationId}:
    get:
      description: Returns a verified location. Planning copies it with fork; this
        row itself is not updated here.
      parameters:
      - description: Location UUID
        format: uuid
        in: path
        name: locationId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Location'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/share.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/share.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/share.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get location
      tags:
      - planning
...
  /trips/{tripId}:
    delete:
      description: Leader only. Members cascade-delete with the trip.
      parameters:
      - description: Trip UUID
        format: uuid
        in: path
        name: tripId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: Trip deleted
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/share.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/share.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/share.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/share.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Delete trip
      tags:
      - trips
    get:
      description: Returns trip info without the invite token. Active members get
        myRole. Private trips are 404 for non-members. Public trips are readable by
        any authenticated user.
      parameters:
      - description: Trip UUID
        format: uuid
        in: path
        name: tripId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/Road-To-Destination-BE_module_trip_model_response.TripResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/share.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/share.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/share.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get trip
      tags:
      - trips
    patch:
      consumes:
      - application/json
      description: Leader or admin. Change name, note, startTime, endTime, and/or
        visibility. Trip type and member limit are not accepted. Status is not updated
        here.
      parameters:
      - description: Trip UUID
        format: uuid
        in: path
        name: tripId
        required: true
        type: string
      - description: Fields to update
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/request.UpdateTripRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/Road-To-Destination-BE_module_trip_model_response.TripResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/share.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/share.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/share.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/share.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update trip
      tags:
      - trips
...
  /trips/{tripId}/fork:
    post:
      consumes:
      - application/json
      description: Any active member forks the trip. The caller is the new leader.
        Other active members are copied as members, minus excludeUserIds. Trip type
        and member limit follow the new trip's tier and cannot be set freely. The
        route graph is not copied.
      parameters:
      - description: Source trip UUID
        format: uuid
        in: path
        name: tripId
        required: true
        type: string
      - description: Optional overrides and members to drop
        in: body
        name: body
        schema:
          $ref: '#/definitions/request.ForkTripRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/Road-To-Destination-BE_module_trip_model_response.TripResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/share.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/share.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/share.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/share.ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/share.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Fork trip
      tags:
      - trips
...
  /trips/{tripId}/invite/{userId}:
    post:
      description: Any active member may invite by user id. Leader or admin invites
        are status invited (the invitee accepts). Member invites are status pending
        until a leader or admin approves. There is no friend check yet.
      parameters:
      - description: Trip UUID
        format: uuid
        in: path
        name: tripId
        required: true
        type: string
      - description: User UUID
        format: uuid
        in: path
        name: userId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.TripMemberResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/share.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/share.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/share.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/share.ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/share.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Invite a user to a trip
      tags:
      - trips
...
  /trips/join/{token}:
    post:
      description: Authenticated caller opens a pending join request for the invite
        token. A leader or admin approves it. Already-active members, existing invites,
        and a full trip return 409.
      parameters:
      - description: Invite token (UUID)
        in: path
        name: token
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.TripMemberResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/share.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/share.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/share.ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/share.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Request to join a trip via invite link
      tags:
      - trips
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
  PlaygroundKey:
    description: Shared secret from GOONG_PLAYGROUND_KEY. Scalar → Authorize.
    in: header
    name: X-Playground-Key
    type: apiKey
swagger: "2.0"